+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # ChangeLog
 
-##  v0.1.0 - v0.1.4 - 20250226
+## v0.1.5 - 20250227
+- 优化 Readme 文件、修复 Toast、转移 Turnstile 展示页面
+
+## v0.1.0 - v0.1.4 - 20250226
 - 移动端兼容
-- 补充 LICENSE、接入 Cloudflare Turnstile
+- 补充 LICENSE、接入 Cloudflare Turnstile、补充 .env.example
 - 上线 uuid、hash、base64、url-codec 等 4 个页面
 - 上线 timestamp、regex、json-formatter、sql-formatter 等 4 个页面
 - init